Support read multi-values from OpenSearch if no codegen triggered (#5015)
* Support read multi-values from OpenSearch


* Fix tests


* remove reflection way


---------

(cherry picked from commit 98516d6e17ad2281f173a7bb0f2d5d1d9232b1ee)
